Big new update to SeekVideoFrame.
Changed queue_page to return error code from calling ogg_sync_pageout.

Created a helper function call BinarySeekFrame, this function uses a binary search to find the frame specified, and if it's a key frame decodes it.  BinarySeekFrame returns the granulepos of the frame it found.

SeekVideoFrame calls BinarySeekFrame for the frame being searched for.  If it's not a keyframe, a second BinarySeekFrame is called to find the keyframe.  Then it keeps decoding till it reaches the target frame.

Also updated UnitTest to seek backwards as a part of it's test.

(SVN Revision 6743 from svn://localhost/TRUNK/)