CameraMPI:  Add audio recording to video capture.  Audio capture runs at the microphone's native 48KHz, mono, signed 16-bit little endian.
NOTE:  This checkin makes Brio dependent on alsa-lib (libasound), which
is not present in Brio/ThirdParty, but rather on the nfsroot via a LinuxDist
package.  This means that Brio is no longer self-contained, and there
might be some build-related heartburn associated with this transition.

(SVN Revision 7135 from svn://localhost/TRUNK/)