microphone: fix input overrun handling
The Alsa documentation says that we should call snd_pcm_recover() when an
overrun occurs.  Note that a pcm_read would return -22 (-EPIPE) and we will
tell Alsa that this is the error we got.  Please see:


This seems to resolve the audio stutters problem with the microphone caused
by doing a pcm_resume in System/Camera/Microphone.cpp RecordCallback() when
the XRUN state is detected.


(SVN Revision 7351 from svn://localhost/TRUNK/)