TTPro 2536.  Adding boolean option to SeekVideoFrame and SeekVideoKeyFrame, bUpdateVideoDisplay.  If set to true, the screen will be updated to the newly seeked position.
This is to get around XMV's need to unpause, seek, and then repause to get screen to update.  The VideoTask was getting confused by the rapid change in state.

Also protecting PutVideoFrame with Mutex, this prevents occasional crashes caused when seek is done while VideoTask is decoding.

(SVN Revision 7388 from svn://localhost/TRUNK/)