CIP 144.  Inlining SetWindowPosition code into RegisterLayer (excluding the layer enable/disable).
The problem is when there are 2 buffers that share the same layer.  Registering the second buffer causes SetWindowPosition to disable the layer.  This is because the second buffer is not enabled yet.  The layer was previously visible from the registration of the first layer.


(SVN Revision 9079 from svn://localhost/TRUNK/)